requirements.txt
# GUI
PySide6>=6.5.0

# 音频处理
pydub>=0.25.1
soundfile>=0.12.1
scipy>=1.10.0
noisereduce>=3.0.0     # 降噪库
pygame>=2.5.0

# ASR
openai-whisper>=20231117
faster-whisper>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0

# 翻译
transformers>=4.35.0
sentencepiece>=0.1.99   # NLLB需要
accelerate>=0.24.0      # 加速推理

# 工具
numpy>=1.24.0
tqdm>=4.65.0

# 系统依赖（非 pip）：必须安装 ffmpeg/ffprobe 并加入 PATH
# Windows 可用：winget install Gyan.FFmpeg